move and refresh
